litton7-proxy/requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
gradio==5.18.0
gradio_client==1.7.2
matplotlib==3.10.0
pillow==11.1.0
